Renamed source_type. Removed unused functions. Corrected upgrade-step.
