+++ src/pkan/dcatapde/harvesting/configure.zcml
@@ -1,7 +1,7 @@
 <configure xmlns="http://namespaces.zope.org/zope"
     i18n_domain="pkan.dcatapde">
 
-  <include package=".source_type"/>
+  <include package=".rdf"/>
 
 
 </configure>
