+++ src/pkan/dcatapde/upgrades.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 """Upgrades."""
-
 from pkan.dcatapde.constants import CT_HARVESTER
 from pkan.dcatapde.constants import CT_HARVESTER_FOLDER
 from pkan.dcatapde.constants import DCAT_CTs
@@ -8,10 +7,13 @@ from plone.app.upgrade.utils import loadMigrationProfile
 from plone.dexterity.factory import DexterityFactory
 from plone.dexterity.interfaces import IDexterityFactory
 from plone.dexterity.interfaces import IDexterityFTI
-from zope.component import getGlobalSiteManager
+from Products.CMFCore.interfaces import ISiteRoot
+from zope.component import getSiteManager
 from zope.component import getUtility
 from zope.component import IFactory
 from zope.component import queryUtility
+
+import transaction
 
 
 def reload_gs_profile(context):
@@ -28,18 +30,24 @@ def remove_defaultfactories(context):
     :param context:
     :return:
     """
-    gsm = getGlobalSiteManager()
+    site = queryUtility(ISiteRoot)
+    if site is None:
+        return
+
+    sm = getSiteManager(site)
     names = DCAT_CTs + [CT_HARVESTER, CT_HARVESTER_FOLDER]
     for name in names:
-        gsm.unregisterUtility(name=name, provided=IDexterityFactory)
+
+        sm.unregisterUtility(name=name, provided=IDexterityFactory)
 
         fti = getUtility(IDexterityFTI, name=name)
 
         factory_utility = queryUtility(IFactory, name=fti.factory)
         if factory_utility is None:
-            gsm.registerUtility(
+            sm.registerUtility(
                 DexterityFactory(name),
                 IFactory,
                 fti.factory,
                 info='plone.dexterity.dynamic',
             )
+    transaction.commit()
